Initial commit
Initial commit of torch_astype. The general approach is hacky, but
simple: we add a .astype method to torch.Tensor that parses string
dtypes. The goal is to improve interoperability with NumPy and to
improve config-file based applications (that use JSON, YAML, etc.).

We plan to extend this to libraries that integrate with PyTorch (such as
bits and bytes).